testing/web-platform/tests/encoding-detection/pl-windows-1250.tentative.html

<!doctype html>
<title>pl windows-1250</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>To jest test kodowania znak�w. W przypadku niekt�rych j�zyk�w, kt�re u�ywaj� znak�w �aci�skich, potrzebujemy wi�cej danych, aby podj�� decyzj�.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "windows-1250", 'Expected windows-1250');
    }, "Check detection result");
      done();
};
</script>
